+++ trunk/ProjectCode/SGVcode/CuiLib/SGVcode/LabColorAnalyzeCluster _veins.cpp
@@ -285,7 +285,7 @@ double *Matrix_W=pMD->Matrix_W_Color;
 {
 	ASSERT(c.A_Color>=-127&&c.A_Color<=128);
 	ASSERT(c.B_color>=-127&&c.B_color<=128);
-	double Length2ORG=sqrtl(pow(c.A_Color,2)+pow(c.B_color,2));
+	float Length2ORG=sqrtl(pow(c.A_Color,2)+pow(c.B_color,2));
 	if (Length2ORG<=CylinderBlackWhite){
 		return true;
 	}else{
@@ -300,9 +300,9 @@ double *Matrix_W=pMD->Matrix_W_Color;
  *
  */
  /*---------------------------------------------------------------------------*/
- double LabColorAnalyzeCluster_Veins::LabDistanceAB(LabColor ci,LabColor cj)
+ float LabColorAnalyzeCluster_Veins::LabDistanceAB(LabColor ci,LabColor cj)
  {
-	 double abdst=sqrtl(pow(ci.A_Color-cj.A_Color,2)+pow(ci.B_color-cj.B_color,2));
+	 float abdst=sqrtl(pow(ci.A_Color-cj.A_Color,2)+pow(ci.B_color-cj.B_color,2));
 	 ASSERT(abdst<=361&&abdst>=0);
 	 return abdst;
  }
@@ -313,9 +313,9 @@ double *Matrix_W=pMD->Matrix_W_Color;
  *
  */
  /*---------------------------------------------------------------------------*/
- double LabColorAnalyzeCluster_Veins::LabDistanceLAB(LabColor ci,LabColor cj)
+ float LabColorAnalyzeCluster_Veins::LabDistanceLAB(LabColor ci,LabColor cj)
  {
-	 double labdst=sqrtl(pow(ci.A_Color-cj.A_Color,2)+pow(ci.B_color-cj.B_color,2)+pow(ci.L_Color-cj.L_Color,2));
+	 float labdst=sqrtl(pow(ci.A_Color-cj.A_Color,2)+pow(ci.B_color-cj.B_color,2)+pow(ci.L_Color-cj.L_Color,2));
 	 ASSERT(labdst<=442&&labdst>=0);
 	 return labdst;
  }
@@ -331,12 +331,12 @@ bool LabColorAnalyzeCluster_Veins::LabIsSameColorInOut(LabColor ci,LabColor cj,b
 	 if (LabDistanceLAB(ci,cj)>=8){
 	 	return false;
 	 }
-	 double DstLab=LabDistanceLAB(ci,cj);
+	 float DstLab=LabDistanceLAB(ci,cj);
 	 if (DstLab<=3){
 		 return true;
 	 }
    if (Enable){
-	   double DstAngle=LabDistanceABangleInOut(ci,cj);
+	   float DstAngle=LabDistanceABangleInOut(ci,cj);
 	   if(DstAngle<=ABInCircleDst){
 		   return true; 
 	   }
@@ -352,12 +352,12 @@ bool LabColorAnalyzeCluster_Veins::LabIsSameColorInOut(LabColor ci,LabColor cj,b
  /*---------------------------------------------------------------------------*/
 bool LabColorAnalyzeCluster_Veins::LabIsSameColorIn(LabColor ci,LabColor cj,bool Enable)
  {
-	 double DstLab=LabDistanceLAB(ci,cj);
+	 float DstLab=LabDistanceLAB(ci,cj);
 	 if (DstLab<=3){
 		 return true;
 	 }
     if (Enable){
-		 double DstAngle=LabDistanceABangleIn(ci,cj);
+		 float DstAngle=LabDistanceABangleIn(ci,cj);
 			 if(DstAngle<=ABInCircleDst){
 	 			return true; 
 			 }
@@ -374,12 +374,12 @@ bool LabColorAnalyzeCluster_Veins::LabIsSameColorIn(LabColor ci,LabColor cj,bool
 /*---------------------------------------------------------------------------*/
 bool LabColorAnalyzeCluster_Veins::LabIsSameColorOut(LabColor ci,LabColor cj,bool Enable)
  {
-	 double DstLab=LabDistanceLAB(ci,cj);
+	 float DstLab=LabDistanceLAB(ci,cj);
 	 if (DstLab<=3){
 		 return true;
 	 }
 	 if (Enable){
-		 double DstAngle=LabDistanceABangleOut(ci,cj);
+		 float DstAngle=LabDistanceABangleOut(ci,cj);
 		 if(DstAngle<=ABOutCircleDst){
 			 return true; 
 		 }
@@ -437,25 +437,25 @@ bool LabColorAnalyzeCluster_Veins::LabIsSameColor(LabColor ci,LabColor cj)
  *
  */
  /*-------------------------------------------------------------------------------------------*/
- double LabColorAnalyzeCluster_Veins::LabDistanceABangleIn(LabColor ci,LabColor cj)
+ float LabColorAnalyzeCluster_Veins::LabDistanceABangleIn(LabColor ci,LabColor cj)
 {
 	ASSERT(ci.A_Color>=-127&&ci.A_Color<=128);
 	ASSERT(cj.B_color>=-127&&cj.B_color<=128);
-	double I_Length=sqrtl(pow(ci.A_Color,2)+pow(ci.B_color,2));
+	float I_Length=sqrtl(pow(ci.A_Color,2)+pow(ci.B_color,2));
 	ASSERT(I_Length!=0);
-	double J_Length=sqrtl(pow(cj.A_Color,2)+pow(cj.B_color,2));
+	float J_Length=sqrtl(pow(cj.A_Color,2)+pow(cj.B_color,2));
 	ASSERT(J_Length!=0);
-	double Vlength=fabs(I_Length-J_Length);
+	float Vlength=fabs(I_Length-J_Length);
 	if (Vlength>=VectorLengthDeltaIn){
 		return 180;
 	}
-	double LDst=fabsl(ci.L_Color-cj.L_Color);
+	float LDst=fabsl(ci.L_Color-cj.L_Color);
 
 	if (LDst>= L_InLightDstDelta){
 		return 180;
 	}
-	double IJ_DotProduct=ci.A_Color*cj.A_Color+ci.B_color*cj.B_color;
-	double CosIJ=IJ_DotProduct/(I_Length*J_Length);
+	float IJ_DotProduct=ci.A_Color*cj.A_Color+ci.B_color*cj.B_color;
+	float CosIJ=IJ_DotProduct/(I_Length*J_Length);
 	if (CosIJ>1){
 		CosIJ-=1E-6F;
 	}
@@ -463,7 +463,7 @@ bool LabColorAnalyzeCluster_Veins::LabIsSameColor(LabColor ci,LabColor cj)
 		CosIJ+=1E-6F;
 	}
 	ASSERT(CosIJ>=-1&&CosIJ<=1);
-	double Angle=acos(CosIJ)*180/CV_PI;
+	float Angle=acos(CosIJ)*180/CV_PI;
 	ASSERT(Angle>=0&&Angle<=181);
 	return Angle;
 }
@@ -473,27 +473,27 @@ bool LabColorAnalyzeCluster_Veins::LabIsSameColor(LabColor ci,LabColor cj)
  *
  */
  /*-------------------------------------------------------------------------------------------*/
- double LabColorAnalyzeCluster_Veins::LabDistanceABangleOut(LabColor ci,LabColor cj)
+ float LabColorAnalyzeCluster_Veins::LabDistanceABangleOut(LabColor ci,LabColor cj)
 {
 	ASSERT(ci.A_Color>=-127&&ci.A_Color<=128);
 	ASSERT(cj.B_color>=-127&&cj.B_color<=128);
-	double I_Length=sqrtl(pow(ci.A_Color,2)+pow(ci.B_color,2));
+	float I_Length=sqrtl(pow(ci.A_Color,2)+pow(ci.B_color,2));
 	ASSERT(I_Length!=0);
-	double J_Length=sqrtl(pow(cj.A_Color,2)+pow(cj.B_color,2));
+	float J_Length=sqrtl(pow(cj.A_Color,2)+pow(cj.B_color,2));
 	ASSERT(J_Length!=0);
-	double Vlength=fabs(I_Length-J_Length);
+	float Vlength=fabs(I_Length-J_Length);
 	if (Vlength>=VectorLengthDeltaOut){
 		return 180;
 	}
-	double LDst=fabsl(ci.L_Color-cj.L_Color);
+	float LDst=fabsl(ci.L_Color-cj.L_Color);
 
 	if (LDst>= L_OutLightDstDelta){
 		return 180;
 	}
-	double IJ_DotProduct=ci.A_Color*cj.A_Color+ci.B_color*cj.B_color;
-	double CosIJ=IJ_DotProduct/(I_Length*J_Length)-1E-6F;
+	float IJ_DotProduct=ci.A_Color*cj.A_Color+ci.B_color*cj.B_color;
+	float CosIJ=IJ_DotProduct/(I_Length*J_Length)-1E-6F;
 	ASSERT(CosIJ>=-1&&CosIJ<=1);
-	double Angle=acos(CosIJ)*180/CV_PI;
+	float Angle=acos(CosIJ)*180/CV_PI;
 	ASSERT(Angle>=0&&Angle<=181);
 	return Angle;
 }
@@ -503,30 +503,30 @@ bool LabColorAnalyzeCluster_Veins::LabIsSameColor(LabColor ci,LabColor cj)
  *
  */
  /*-------------------------------------------------------------------------------------------*/
- double LabColorAnalyzeCluster_Veins::LabDistanceABangleInOut(LabColor ci,LabColor cj)
+ float LabColorAnalyzeCluster_Veins::LabDistanceABangleInOut(LabColor ci,LabColor cj)
  {
 
 	ASSERT(ci.A_Color>=-127&&ci.A_Color<=128);
 	ASSERT(cj.B_color>=-127&&cj.B_color<=128);
-	double I_Length=sqrtl(pow(ci.A_Color,2)+pow(ci.B_color,2));
+	float I_Length=sqrtl(pow(ci.A_Color,2)+pow(ci.B_color,2));
 	ASSERT(I_Length!=0);
-	double J_Length=sqrtl(pow(cj.A_Color,2)+pow(cj.B_color,2));
+	float J_Length=sqrtl(pow(cj.A_Color,2)+pow(cj.B_color,2));
 	ASSERT(J_Length!=0);
-	double Vlength=fabs(I_Length-J_Length);
+	float Vlength=fabs(I_Length-J_Length);
 	if (Vlength>=VectorLengthDeltaOut){
 		return 180;
 	}
 	
-	double LDst=fabsl(ci.L_Color-cj.L_Color);
+	float LDst=fabsl(ci.L_Color-cj.L_Color);
 	
 		if (LDst>= L_OutLightDstDelta){
 			return 180;
 		}
 	
-	double IJ_DotProduct=ci.A_Color*cj.A_Color+ci.B_color*cj.B_color;
-	double CosIJ=IJ_DotProduct/(I_Length*J_Length)-1E-6F;
+	float IJ_DotProduct=ci.A_Color*cj.A_Color+ci.B_color*cj.B_color;
+	float CosIJ=IJ_DotProduct/(I_Length*J_Length)-1E-6F;
 	ASSERT(CosIJ>=-1&&CosIJ<=1);
-	double Angle=acos(CosIJ)*180/CV_PI;
+	float Angle=acos(CosIJ)*180/CV_PI;
 	ASSERT(Angle>=0&&Angle<=181);
 	return Angle;
  }
@@ -536,19 +536,19 @@ bool LabColorAnalyzeCluster_Veins::LabIsSameColor(LabColor ci,LabColor cj)
 *
 */
 /*-------------------------------------------------------------------------------------------*/
- double LabColorAnalyzeCluster_Veins::LabDistanceLABangle(LabColor ci,LabColor cj)
+ float LabColorAnalyzeCluster_Veins::LabDistanceLABangle(LabColor ci,LabColor cj)
  {
 	ASSERT(ci.A_Color>=-127&&ci.A_Color<=128);
 	ASSERT(cj.B_color>=-127&&cj.B_color<=128);
-	double I_Length=sqrtl(pow(ci.A_Color,2)+pow(ci.B_color,2)+pow(ci.L_Color,2));
+	float I_Length=sqrtl(pow(ci.A_Color,2)+pow(ci.B_color,2)+pow(ci.L_Color,2));
 	ASSERT(I_Length!=0);
-	double J_Length=sqrtl(pow(cj.A_Color,2)+pow(cj.B_color,2)+pow(cj.L_Color,2));
+	float J_Length=sqrtl(pow(cj.A_Color,2)+pow(cj.B_color,2)+pow(cj.L_Color,2));
 	ASSERT(J_Length!=0);
 	
-	double IJ_DotProduct=ci.A_Color*cj.A_Color+ci.B_color*cj.B_color+ci.L_Color*cj.L_Color;
-	double CosIJ=IJ_DotProduct/(I_Length*J_Length);
+	float IJ_DotProduct=ci.A_Color*cj.A_Color+ci.B_color*cj.B_color+ci.L_Color*cj.L_Color;
+	float CosIJ=IJ_DotProduct/(I_Length*J_Length);
 	ASSERT(CosIJ>=-1&&CosIJ<=1);
-	double Angle=acos(CosIJ)*180/CV_PI;
+	float Angle=acos(CosIJ)*180/CV_PI;
 	return  Angle;
  }
  /*-------------------------------------------------------------------------------------------*/
@@ -585,10 +585,10 @@ bool LabColorAnalyzeCluster_Veins::LabIsSameColor(LabColor ci,LabColor cj)
 	 double *Matrix_W_Vein,
 	 double *Matrix_W_Multi,
 	 int spcount,
-	 double Vein_threshold_SkyV,
-	 double Vein_threshold_GND,
-	 double Color_threshold_SkyV,
-	 double Color_threshold_GND,
+	 float Vein_threshold_SkyV,
+	 float Vein_threshold_GND,
+	 float Color_threshold_SkyV,
+	 float Color_threshold_GND,
 	 SP_PROPERTY *p_SpProperty)
  {
 #if 0
